# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "AttemptRecord": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deadline_at": {
            "format": "date-time",
            "title": "Deadline At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "project_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "proposal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProposalRecord"
              },
              {
                "type": "null"
              }
            ]
          },
          "reserved_tokens": {
            "title": "Reserved Tokens",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "status",
          "reserved_tokens",
          "input_tokens",
          "output_tokens",
          "error_code",
          "created_at",
          "deadline_at"
        ],
        "title": "AttemptRecord",
        "type": "object"
      },
...
      "ContextPlan": {
        "additionalProperties": false,
        "properties": {
          "acceptance_criteria": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 16,
            "minItems": 1,
            "title": "Acceptance Criteria",
            "type": "array"
          },
          "assumptions": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 10,
            "title": "Assumptions",
            "type": "array"
          },
          "out_of_scope": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 10,
            "title": "Out Of Scope",
            "type": "array"
          },
          "plan": {
            "$ref": "#/components/schemas/ProjectPlan"
          },
          "scope": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 16,
            "minItems": 1,
            "title": "Scope",
            "type": "array"
          }
        },
        "required": [
          "plan",
          "scope",
          "assumptions",
          "acceptance_criteria",
          "out_of_scope"
        ],
        "title": "ContextPlan",
        "type": "object"
      },
      "CoreFeature": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "description"
        ],
        "title": "CoreFeature",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ImplementationMilestone": {
        "additionalProperties": false,
        "properties": {
          "deliverables": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Deliverables",
            "type": "array"
          },
          "title": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "deliverables"
        ],
        "title": "ImplementationMilestone",
        "type": "object"
      },
      "PlanInput": {
        "additionalProperties": false,
        "properties": {
          "base_brain_revision_id": {
            "format": "uuid",
            "title": "Base Brain Revision Id",
            "type": "string"
          },
          "base_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Version Id"
          },
          "kind": {
            "enum": [
              "initial",
              "change"
            ],
            "title": "Kind",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "request_id",
          "base_brain_revision_id",
          "base_version_id"
        ],
        "title": "PlanInput",
        "type": "object"
      },
...
      "ProjectPlan": {
        "additionalProperties": false,
        "properties": {
          "core_features": {
            "items": {
              "$ref": "#/components/schemas/CoreFeature"
            },
            "maxItems": 16,
            "minItems": 1,
            "title": "Core Features",
            "type": "array"
          },
          "implementation_milestones": {
            "items": {
              "$ref": "#/components/schemas/ImplementationMilestone"
            },
            "maxItems": 12,
            "minItems": 1,
            "title": "Implementation Milestones",
            "type": "array"
          },
          "product_summary": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Product Summary",
            "type": "string"
          },
          "project_title": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Project Title",
            "type": "string"
          },
          "recommended_stack": {
            "$ref": "#/components/schemas/RecommendedStack"
          },
          "target_users": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Target Users",
            "type": "array"
          }
        },
        "required": [
          "project_title",
          "product_summary",
          "target_users",
          "core_features",
          "recommended_stack",
          "implementation_milestones"
        ],
        "title": "ProjectPlan",
        "type": "object"
      },
...
      "ProposalList": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProposalRecord"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ProposalList",
        "type": "object"
      },
      "ProposalRecord": {
        "additionalProperties": false,
        "properties": {
          "attempt_id": {
            "format": "uuid",
            "title": "Attempt Id",
            "type": "string"
          },
          "brain_revision_id": {
            "format": "uuid",
            "title": "Brain Revision Id",
            "type": "string"
          },
          "content": {
            "$ref": "#/components/schemas/ContextPlan"
          },
          "context": {
            "additionalProperties": true,
            "title": "Context",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_context": {
            "title": "Current Context",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "initial",
              "change"
            ],
            "title": "Kind",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "reviewed": {
            "title": "Reviewed",
            "type": "boolean"
          },
          "version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Id"
          }
        },
        "required": [
          "id",
          "attempt_id",
          "project_id",
          "request_id",
          "brain_revision_id",
          "version_id",
          "kind",
          "content",
          "context",
          "provider",
          "model",
          "created_at",
          "current_context",
          "reviewed"
        ],
        "title": "ProposalRecord",
        "type": "object"
      },
...
      "RecommendedStack": {
        "additionalProperties": false,
        "properties": {
          "backend": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Backend",
            "type": "string"
          },
          "database": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Database",
            "type": "string"
          },
          "frontend": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Frontend",
            "type": "string"
          },
          "rationale": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          }
        },
        "required": [
          "frontend",
          "backend",
          "database",
          "rationale"
        ],
        "title": "RecommendedStack",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AuthenticatedBearer": {
        "description": "Verified provider API token plus server-owned session in OIDC mode; private local credential only in explicit development mode.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "F01 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/projects/{project_id}/planning/attempts": {
      "post": {
        "operationId": "plan_v1_projects__project_id__planning_attempts_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "AuthenticatedBearer": []
          }
        ],
        "summary": "Plan",
        "tags": [
          "Contextual planning"
        ]
      }
    },
    "/v1/projects/{project_id}/planning/attempts/{attempt_id}": {
      "get": {
        "operationId": "attempt_v1_projects__project_id__planning_attempts__attempt_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "AuthenticatedBearer": []
          }
        ],
        "summary": "Attempt",
        "tags": [
          "Contextual planning"
        ]
      }
    },
    "/v1/projects/{project_id}/planning/attempts/{attempt_id}/cancel": {
      "post": {
        "operationId": "cancel_v1_projects__project_id__planning_attempts__attempt_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "AuthenticatedBearer": []
          }
        ],
        "summary": "Cancel",
        "tags": [
          "Contextual planning"
        ]
      }
    },
    "/v1/projects/{project_id}/planning/proposals": {
      "get": {
        "operationId": "list_proposals_v1_projects__project_id__planning_proposals_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalList"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "AuthenticatedBearer": []
          }
        ],
        "summary": "List Proposals",
        "tags": [
          "Contextual planning"
        ]
      }
    },
    "/v1/projects/{project_id}/planning/proposals/{proposal_id}/review": {
      "post": {
        "operationId": "review_v1_projects__project_id__planning_proposals__proposal_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "AuthenticatedBearer": []
          }
        ],
        "summary": "Review",
        "tags": [
          "Contextual planning"
        ]
      }
    },
...
  }
}